# packages/sdk/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/support/settings": {
      "get": {
        "description": "Retrieve support configuration for merchant. Includes FAQ items and routing rules.",
        "operationId": "Support_getSettings_",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Support settings retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "merchantId": "merch_123",
                    "faqItems": [],
                    "enabled": true,
                    "createdAt": "2026-08-10T00:00:00Z"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing support:read scope"
          },
          "default": {
            "description": "RFC 7807 error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "console_session": []
          },
          {
            "service_api_key": []
          },
          {
            "legacy_api_key": []
          }
        ],
        "summary": "Get support settings",
        "tags": [
          "Support"
        ]
      },
      "put": {
        "description": "Update support configuration including FAQ items. Requires idempotency key.",
        "operationId": "Support_updateSettings_",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "description": "Unique key for safely retrying this mutation. Reuse with a different payload returns 409.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 255
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupportSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Support settings updated"
          },
          "400": {
            "description": "Invalid settings body"
          },
          "403": {
            "description": "Missing support:write scope"
          },
          "default": {
            "description": "RFC 7807 error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "console_session": []
          },
          {
            "service_api_key": []
          },
          {
            "legacy_api_key": []
          }
        ],
        "summary": "Update support settings",
        "tags": [
          "Support"
        ]
      }
    },
    "/v1/support/tickets": {
      "get": {
        "description": "List support tickets with pagination. Supports filtering by status and cursor-based pagination.",
        "operationId": "Support_getTickets",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Support tickets retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": [
                      {
                        "id": "ticket_123",
                        "status": "open",
                        "createdAt": "2026-08-10T00:00:00Z"
                      }
                    ],
                    "next_cursor": "cursor_next",
                    "has_more": true
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing support:read scope"
          },
          "default": {
            "description": "RFC 7807 error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "console_session": []
          },
          {
            "service_api_key": []
          },
          {
            "legacy_api_key": []
          }
        ],
        "summary": "List support tickets",
        "tags": [
          "Support"
        ]
      },
      "post": {
        "description": "Create a new support ticket associated with a session. Requires idempotency key.",
        "operationId": "Support_createTicket_",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "description": "Unique key for safely retrying this mutation. Reuse with a different payload returns 409.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 255
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Support ticket created"
          },
          "400": {
            "description": "Invalid ticket body"
          },
          "403": {
            "description": "Missing support:write scope"
          },
          "default": {
            "description": "RFC 7807 error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "console_session": []
          },
          {
            "service_api_key": []
          },
          {
            "legacy_api_key": []
          }
        ],
        "summary": "Create a support ticket",
        "tags": [
          "Support"
        ]
      }
    },
    "/v1/support/tickets/{ticketId}": {
      "patch": {
        "description": "Transition a support ticket to a new status (open, in_progress, resolved, closed). Requires idempotency key.",
        "operationId": "Support_updateTicket",
        "parameters": [
          {
            "name": "ticketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "description": "Unique key for safely retrying this mutation. Reuse with a different payload returns 409.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 255
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupportTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket status updated"
          },
          "400": {
            "description": "Invalid status transition"
          },
          "403": {
            "description": "Missing support:write scope"
          },
          "404": {
            "description": "Ticket not found"
          },
          "default": {
            "description": "RFC 7807 error response.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "console_session": []
          },
          {
            "service_api_key": []
          },
          {
            "legacy_api_key": []
          }
        ],
        "summary": "Update support ticket status",
        "tags": [
          "Support"
        ]
      }
    },
...
  },
  "info": {
    "title": "AACP Integration API",
    "description": "Public API for merchant configuration, checkout integrations and operational workflows. All tenant access is derived from the authenticated credential.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://sandbox-api.aacp.dev",
      "description": "Sandbox"
    },
    {
      "url": "https://api.aacp.dev",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "service_api_key": {
        "scheme": "bearer",
        "bearerFormat": "aacp_test_* or aacp_live_*",
        "type": "http",
        "description": "Service API key issued by the merchant console."
      },
      "console_session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aacp_access_token",
        "description": "Short-lived HttpOnly merchant console session."
      },
      "legacy_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-aacp-api-key",
        "description": "Deprecated compatibility header. Prefer Bearer authentication."
      }
    },
    "schemas": {
...
      "SupportFaqItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "question",
          "answer"
        ]
      },
      "UpdateSupportSettingsDto": {
        "type": "object",
        "properties": {
          "faqItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupportFaqItemDto"
            }
          }
        },
        "required": [
          "faqItems"
        ]
      },
      "CreateSupportTicketDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "UpdateSupportTicketDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "resolved",
              "closed"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "correlation_id"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "code": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "correlation_id": {
            "type": "string"
          }
        }
      }
    }
  }
}